Split djangify.py to processing_utils.py and djangify.py Remove the main() from djangify.py
