+++ djangify.egg-info/SOURCES.txt
@@ -3,6 +3,7 @@ setup.cfg
 setup.py
 djangify/__init__.py
 djangify/djangify.py
+djangify/processing_utils.py
 djangify.egg-info/PKG-INFO
 djangify.egg-info/SOURCES.txt
 djangify.egg-info/dependency_links.txt
