+++ djangify/djangify.py
@@ -26,9 +26,10 @@ optional arguments:
 """
 
 import os
-import re
 import argparse
 import logging
+
+from .processing_utils import process_line
 
 # Set default logging level to INFO
 logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.INFO)
@@ -49,164 +50,7 @@ def display_path_info():
     logging.info("Directory name is : " + foldername)
 
 
-def check_line(line: str):
-    """
-    A funtion that checks if a string passed to this function contains key
-    words that need to be processed further
-
-    Parameters
-    ----------
-    line : str\n
-            a string that needs to be checked if it contains key words
-
-    Returns
-    -------
-    list\n
-            a list of words found in the string 'line', if the word is
-            a keyword,then instead of only the word, a tuple in the
-            form of (True, word) is added
-    """
-
-    key_words = ['src', 'href', 'url']
-    out = list()
-    for word in key_words:
-        if line.__contains__(word):
-            out.append((True, word))
-
-    # Check if output list is not empty
-    if len(out) == 0:
-        # If list is empty return None
-        return None
-    else:
-        return out
-
-
-def contains_url(line: str):
-    """
-    Checks if the line contains any URLs
-
-    Parameters
-    ----------
-    line : str\n
-            The string that needs to be checked if it countains URLs
-
-    Returns
-    -------
-    bool\n
-            True if it contains URL and False if no URL in 'line'
-    """
-
-    URL = r"(http|ftp|https)://([\w_-]+(?:(?:\.[\w_-]+)+))" \
-          r"([\w.,@?^=%&:/~+#-]*[\w@?^=%&/~+#-])?"
-    if re.match(URL, line):
-        return True
-    else:
-        return False
-
-
-def get_index(line: str, word: str):
-    """
-    Get the starting and ending index of a word in a given string
-
-    Parameters
-    ----------
-    line : str\n
-            A string containing the 'word' from which indexes are
-            to be extracted from
-    word : str\n
-            A string that need to be found in the 'line', and need indexes
-            extracted
-
-    Returns
-    -------
-    tuple\n
-        A tuple of the form (start, end), where start and end are the starting
-        and ending indexes of the 'word' in the given 'line'
-    """
-
-    index = line.find(word)
-
-    if word in ['url']:
-        start = (index + len(word) + 2)
-        quote = line[start - 1]
-        if quote not in ['\'', '"']:
-            start = (index + len(word) + 1)
-            quote = line[start - 1]
-            if quote == '(':
-                end = line.find(')', start)
-            else:
-                end = line.find(quote, start)
-        else:
-            end = line.find(quote, start)
-    else:
-        start = (index + len(word) + 2)
-        quote = line[start - 1]
-        end = line.find(quote, start)
-
-    return start, end
-
-
-def djangify(line: str):
-    """
-    Translates the string passed to the function to Django compatible HTML
-
-    Parameters
-    ----------
-    line : str\n
-            A string that contains native HTML that need translation
-
-    Returns
-    -------
-    str\n
-            Translated HTML that is Django compatible
-    """
-
-    global APP_NAME
-
-    # Don't change the contents of the line if it contails a URL that links
-    # outside content. Ex. www.example.com/webpage.html
-    if contains_url(line):
-        return line
-    # Don't change the line if it contains placeholder URL like '#'
-    if line == '#':
-        return line
-    # If line links to an internal file, make it Django compatible by loading
-    # from static directory appended with APP_NAME
-    return " {% static '" + APP_NAME + line + "' %} "
-
-
-def process_line(line: str):
-    """
-    Processes the line (string) of text passed in as parameter into django
-    compatible HTML by calling the djangify(...) function.
-
-    Parameters
-    ----------
-    line : str\n
-            A line (string) of HTML, that is yet to be made django compatible
-
-    Returns
-    -------
-    str\n
-            Translated HTML that is Django compatible
-    """
-
-    # Converts line into a list of words, using checkLine()
-    instances = check_line(line)
-
-    buffer = line
-
-    if instances:
-        for instance in instances:
-            index = get_index(buffer, instance[1])
-            out = djangify(buffer[index[0]: index[1]])
-            text = buffer[: index[0]] + out + buffer[index[1]:]
-            buffer = text
-
-    return buffer
-
-
-def process_file(directory: str, filepath: str, fname: str):
+def process_file(directory: str, filepath: str, fname: str, app_name: str = APP_NAME):
     """
     Prcocesses the file passed in as parameter, translating it into django
     friendly HTML.
@@ -221,6 +65,8 @@ def process_file(directory: str, filepath: str, fname: str):
             path of the file that is to be translated
     fname : fname\n
             name of the file to be translated
+    app_name: str\n
+            name of the application to be used, defaults to none
     """
 
     fname = fname.split(".")[0]
@@ -242,7 +88,7 @@ def process_file(directory: str, filepath: str, fname: str):
             cnt = 1
             while line:
                 # process the line extracted
-                temp = process_line(line)
+                temp = process_line(line, app_name)
                 line = fp.readline()
                 cnt += 1
                 # write the processed line to the newly created file
@@ -324,15 +170,14 @@ def main():
         os.mkdir(os.path.join(directory, "Modified_files"))
 
     if files:
+        print("here")
         for file in files:
             process_file(directory, directory + "/" + file, file)
 
     else:
+        print("no files found")
         # If no file was passed in as input, then extract all files in the
         # directory passed in, with extension '.html'
         for file in os.listdir(directory):
             if file.endswith(".html"):
                 process_file(directory, directory + "/" + file, file)
-
-
-main()
